portfolio-intake/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-aiplatform
google-auth
gspread
python-dotenv
pypdf
